requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
flask-jwt-extended==4.3.1
bcrypt==3.2.0
python-dateutil==2.8.2
gunicorn==20.1.0
dnspython==2.1.0
